x scale now shown correctly

+++ measureFI.py
@@ -1,18 +1,32 @@
-import numpy as np
-import os
-import sys
-n=3
-os.remove("recfolder/FI.npy")
-data=np.zeros((n),dtype="object")
-np.save("recfolder/FI",data)
-cells=["pyr","bas","olm"]
-for i in range(n):
-    I=i*.125
-    print("python2 my_mosinit.py "+str(I)+" "+cells[i])
-    os.system("python2 my_mosinit.py "+str(I)+" "+cells[i])#script saves in .npy file the value
     
-    #print("args:",sys.argv[1])
-#dat=np.load("recfolder/FI.npy")
+
+
+n=10
+stepsize=[0.25,0.25,1]
+if __name__ == "__main__":
+    import numpy as np
+    import os
+    import sys
+
+    if os.path.exists("recfolder/FI.npy"):
+        os.remove("recfolder/FI.npy")
+    cells=["pyr","bas","olm"]
+
+    Data=np.zeros((len(cells)),dtype="object")
+    np.save("recfolder/FI",Data)
+    for j in range(len(cells)):
+        Data=np.load("recfolder/FI.npy",allow_pickle=True)
+        data=np.zeros((n),dtype="object")
+        Data[j]=data
+        np.save("recfolder/FI",Data)
+        for i in range(n):
+            I=i*stepsize[j]
+            commandstring="python2 my_mosinit.py "+str(I)+" "+cells[j]
+            print(commandstring)
+            os.system(commandstring)#script saves in .npy file the value
+        
+        #print("args:",sys.argv[1])
+    #dat=np.load("recfolder/FI.npy")
